fase-3/requirements.txt
opencv-python
pandas
numpy
joblib
tensorflow
scikit-learn
argparse
loguru
flask
requests
pillow